调整文件

+++ start/start/script/img-write-osx.sh
@@ -2,9 +2,20 @@ if [ -f "disk1.vhd" ]; then
     mv disk1.vhd disk1.dmg
 fi
 
+if [ ! -f "disk1.dmg" ]; then
+    echo "找不到disk1.vhd，请从课程提供的百度网盘链接上下载放到image目录下"
+    exit
+fi
+
 if [ -f "disk2.vhd" ]; then
     mv disk2.vhd disk2.dmg
 fi
+
+if [ ! -f "disk2.dmg" ]; then
+    echo "找不到disk2.vhd，请从课程提供的百度网盘链接上下载放到image目录下"
+    exit
+fi
+
 
 export DISK1_NAME=disk1.dmg
 
@@ -26,7 +37,7 @@ dd if=boot.bin of=$DISK1_NAME bs=512 conv=notrunc count=1
 # export TARGET_PATH=mp
 # rm $TARGET_PATH
 # hdiutil attach $DISK2_NAME -mountpoint $TARGET_PATH
-# cp -v init.elf $TARGET_PATH
+# cp -v init.elf $TARGET_PATH/init
 # cp -v shell.elf $TARGET_PATH
-# cp -v loop $TARGET_PATH
+# cp -v loop.elf $TARGET_PATH/loop
 # hdiutil unmount $TARGET_PATH -verbose
